+++ project_toki/grammar_parser.py
@@ -1,5 +1,3 @@
-from typing import List
-
 import lark
 
 from project_toki.grammar import Grammar
@@ -17,7 +15,7 @@ class GrammarParser:
     - leaf "WS" and leaves with name starting with "PUNCT_" will be treated as punctuation
     - leaves of the grammar tree that are not punctuation must be named after a PartOfSpeech
     - when node name contains "__" then only the part before "__" will appear in the output tree
-    - when node name starts with "p_" then this node will be removed and its parent will adopt its children
+    - when node name starts with "_" then this node will be removed and its parent will adopt its children
     """
 
     _PARSER: lark.Lark = lark.Lark(
@@ -32,7 +30,7 @@ class GrammarParser:
         This method parses input text into a grammatical tree.
         """
         return GrammarParser._tree_to_phrases(
-            GrammarParser._squash_private_nodes(GrammarParser._PARSER.parse(text)),
+            GrammarParser._remove_whitespaces(GrammarParser._PARSER.parse(text)),
         )
 
     @staticmethod
@@ -41,7 +39,7 @@ class GrammarParser:
         This method recursively converts lark tree to phrases and words.
         """
         if isinstance(tree, lark.lexer.Token):
-            if tree.type.startswith("PUNCT_") or tree.type == "WS":
+            if tree.type.startswith("PUNCT_"):
                 return Phrase(
                     name=Punctuation(
                         text=tree.value,
@@ -65,28 +63,21 @@ class GrammarParser:
             )
 
     @staticmethod
-    def _squash_private_nodes(
+    def _remove_whitespaces(
         tree: lark.lexer.Token | lark.tree.Tree,
     ) -> lark.lexer.Token | lark.tree.Tree:
         """
-        This method recursively removes private nodes while retaining its children.
+        This method removes whitespaces from the tree.
         """
         if isinstance(tree, lark.lexer.Token):
             return tree
         else:
             new_children: list[lark.lexer.Token | lark.tree.Tree] = []
-            found_any_private_nodes: bool = False
             for child in tree.children:
                 if isinstance(child, lark.lexer.Token):
-                    new_children.append(child)
-                elif child.data.startswith("p_"):
-                    new_children += [
-                        GrammarParser._squash_private_nodes(x) for x in child.children
-                    ]
-                    found_any_private_nodes = True
+                    if not child.type.startswith("WS"):
+                        new_children.append(child)
                 else:
-                    new_children.append(GrammarParser._squash_private_nodes(child))
+                    new_children.append(GrammarParser._remove_whitespaces(child))
             tree.children = new_children
-            if found_any_private_nodes:
-                tree = GrammarParser._squash_private_nodes(tree)
             return tree